0512/1516 최신 사항 반영
