+++ src/main/java/com/grepp/nbe562team04/app/controller/api/GoalCompanyApiController.java
@@ -1,14 +1,12 @@
-package com.grepp.nbe562team04.goalcompany.controller;
+package com.grepp.nbe562team04.app.controller.api;
 
-import com.grepp.nbe562team04.goalcompany.dto.GoalCompanyRequestDto;
-import com.grepp.nbe562team04.goalcompany.entity.GoalCompany;
-import com.grepp.nbe562team04.goalcompany.repository.GoalCompanyRepository;
-import com.grepp.nbe562team04.user.entity.User;
-import com.grepp.nbe562team04.user.repository.UserRepository;
+import com.grepp.nbe562team04.model.goalcompany.entity.GoalCompany;
+import com.grepp.nbe562team04.model.goalcompany.repository.GoalCompanyRepository;
+import com.grepp.nbe562team04.model.user.entity.User;
+import com.grepp.nbe562team04.model.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
