+++ src/main/java/com/grepp/nbe562team04/app/controller/web/dashboard/DashboardViewController.java
@@ -1,4 +1,4 @@
-package com.grepp.nbe562team04.dashboard.controller;
+package com.grepp.nbe562team04.app.controller.web.dashboard;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
