+++ src/main/java/com/grepp/nbe562team04/model/goalcompany/repository/GoalCompanyRepository.java
@@ -1,6 +1,6 @@
-package com.grepp.nbe562team04.goalcompany.repository;
+package com.grepp.nbe562team04.model.goalcompany.repository;
 
-import com.grepp.nbe562team04.goalcompany.entity.GoalCompany;
+import com.grepp.nbe562team04.model.goalcompany.entity.GoalCompany;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GoalCompanyRepository extends JpaRepository<GoalCompany, Long> {
