+++ src/main/java/com/grepp/nbe562team04/model/todo/entity/Todo.java
@@ -1,41 +1,35 @@
-package com.grepp.nbe562team04.goal.entity;
+package com.grepp.nbe562team04.model.todo.entity;
 
-import com.grepp.nbe562team04.goalcompany.entity.GoalCompany;
+import com.grepp.nbe562team04.model.goalcompany.entity.GoalCompany;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "goal")
+@Table(name = "todos_table")
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class Goal {
+public class Todo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "goal_id")
-    private Long goalId;
+    @Column(name = "todo_id")
+    private Long todoId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "company_id")
     private GoalCompany goalCompany;
 
-    @Column(nullable = false)
-    private String title;
+    private String content;
 
     @Column(name = "start_date")
     private LocalDateTime startDate;
 
     @Column(name = "end_date")
     private LocalDateTime endDate;
-
-    private String status;
-
-    @Column(name = "created_at")
-    private LocalDateTime createdAt;
 
     @Column(name = "is_done")
     private Boolean isDone;
